refactor: moves square-related function from board utils to square utils
